requirements_RPI.txt
###### Requirements for running on an RPI ######
RPi.GPIO